Conflicts:
	admin/client/js/es5/all.min.js
	admin/client/js/es5/all.min.js.map
	admin/client/js/es5/controllers/collections.js.map
	admin/client/js/es5/controllers/createcollection.js.map
	admin/client/js/es5/directives/common/swexpandablerecord.js.map
	admin/client/js/es5/directives/common/swlistingdisplay.js.map
	admin/client/js/es5/directives/content/swcontentlist.js.map
	admin/client/js/es5/directives/orderitem/sworderitems.js.map
	admin/client/js/es5/directives/productBundleGroup/swproductbundlegroup.js.map
	admin/client/js/es5/modules/ngslatwall.js.map
	admin/client/js/es5/modules/ngslatwallmodel.js.map
	admin/client/js/es5/modules/slatwalladmin.js
	admin/client/js/es5/modules/slatwalladmin.js.map
	admin/client/js/es5/services/collectionconfigservice.js.map
	admin/client/js/es5/services/observerservice.js.map
	admin/client/js/es5/services/selectionservice.js.map
	admin/client/js/es5/services/slatwallInterceptor.js.map
	admin/client/js/es5/services/utilityservice.js.map
	admin/client/js/es6/controllers/collections.js.map
	admin/client/js/es6/controllers/createcollection.js.map
	admin/client/js/es6/directives/common/swexpandablerecord.js.map
	admin/client/js/es6/directives/common/swlistingdisplay.js.map
	admin/client/js/es6/directives/content/swcontentlist.js.map
	admin/client/js/es6/directives/orderitem/sworderitems.js.map
	admin/client/js/es6/directives/productBundleGroup/swproductbundlegroup.js.map
	admin/client/js/es6/modules/ngslatwall.js.map
	admin/client/js/es6/modules/ngslatwallmodel.js.map
	admin/client/js/es6/modules/slatwalladmin.js
	admin/client/js/es6/modules/slatwalladmin.js.map
	admin/client/js/es6/services/collectionconfigservice.js.map
	admin/client/js/es6/services/observerservice.js.map
	admin/client/js/es6/services/selectionservice.js.map
	admin/client/js/es6/services/slatwallInterceptor.js.map
	admin/client/js/es6/services/utilityservice.js.map
	admin/client/ts/modules/slatwalladmin.ts
	client/typings/slatwallTypescript.d.ts

+++ model/service/AppService.cfc
@@ -67,11 +67,13 @@ component extends="HibachiService" accessors="true" output="false" {
 	}
 	
 	public void function updateCMSApp(required app){
-		getService("hibachiUtilityService").copyFile(
-			source= replacenocase(getSkeletonAppPath(),'\','/','all') & '/Application.cfc', 
-			destination=replacenocase(arguments.app.getAppPath(),'\','/','all') & '/Application.cfc', 
-			overwrite=true
-		);
+		if(directoryExists(replacenocase(arguments.app.getAppPath(),'\','/','all'))){
+			getService("hibachiUtilityService").copyFile(
+				source= replacenocase(getSkeletonAppPath(),'\','/','all') & '/Application.cfc', 
+				destination=replacenocase(arguments.app.getAppPath(),'\','/','all') & '/Application.cfc', 
+				overwrite=true
+			);
+		}
 	}
 	
 	public string function getSkeletonAppPath(){
